ajustes para jpa de camara y location

+++ Backend/Java/src/main/java/edu/ucaldas/hackathon/DTOs/camara/CreateCamaraDTO.java
@@ -1,9 +1,11 @@
 package edu.ucaldas.hackathon.DTOs.camara;
 
+import java.math.BigDecimal;
+
 public record CreateCamaraDTO(
     String name,
-    double angleXY,
-    double angleXZ,
+    BigDecimal angleXY,
+    BigDecimal angleXZ,
     CreateLocationDTO location
 
 ) {
