+++ Backend/Java/src/main/java/edu/ucaldas/hackathon/DTOs/camara/GetCamaraDTO.java
@@ -1,10 +1,12 @@
 package edu.ucaldas.hackathon.DTOs.camara;
 
+import java.math.BigDecimal;
+
 public record GetCamaraDTO(
     String id,
     String name,
-    double angleXY,
-    double angleXZ,
+    BigDecimal angleXY,
+    BigDecimal angleXZ,
     GetLocationDTO location
 ) {
