+++ Backend/Java/src/main/java/edu/ucaldas/hackathon/DTOs/camara/UpdateCamaraDTO.java
@@ -1,9 +1,11 @@
 package edu.ucaldas.hackathon.DTOs.camara;
 
+import java.math.BigDecimal;
+
 public record UpdateCamaraDTO(
     String name,
-    double angleXY,
-    double angleXZ,
+    BigDecimal angleXY,
+    BigDecimal angleXZ,
     UpdateLocationDTO location
 ) {
